requirements.txt
aiodns==3.2.0
aiohappyeyeballs==2.4.2
aiohttp==3.10.8
aiosignal==1.3.1
alembic==1.13.3
altair==5.4.1
antlr4-python3-runtime==4.9.3
attrs==24.2.0
blinker==1.8.2
build==1.2.2
cachetools==5.5.0
ccxt==4.4.9
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
colorlog==6.8.2
contourpy==1.3.0
cryptography==43.0.1
cycler==0.12.1
dateparser==1.2.0
filelock==3.16.1
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.1
gym==0.26.2
gym-notices==0.0.8
hydra-core==1.3.2
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
lightning-utilities==0.11.7
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
narwhals==1.8.4
networkx==3.3
numpy==2.1.1
omegaconf==2.3.0
optuna==4.0.0
packaging==24.1
pandas==2.2.3
pillow==10.4.0
pip-tools==7.4.1
plotly==5.24.1
pluggy==1.5.0
protobuf==5.28.2
pyarrow==17.0.0
pycares==4.4.0
pycparser==2.22
pycryptodome==3.20.0
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.4
pyproject_hooks==1.2.0
pytest==8.3.3
python-binance==1.0.19
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytorch-lightning==2.4.0
pytz==2024.2
PyYAML==6.0.2
ray==2.37.0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rich==13.8.1
rpds-py==0.20.0
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
setuptools==75.1.0
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.35
streamlit==1.38.0
sympy==1.13.3
ta==0.11.0
tenacity==8.5.0
threadpoolctl==3.5.0
toml==0.10.2
torch==2.4.1
torchmetrics==1.4.2
tornado==6.4.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
ujson==5.10.0
urllib3==2.2.3
watchdog==4.0.2
websocket-client==1.8.0
websockets==13.1
wheel==0.44.0
yarl==1.13.1
